fixed windows script

+++ setup.ps1
@@ -1,47 +1,55 @@
 $REPO_URL = "https://github.com/evanamedi/SpotifyUnwrapped.git"
 $REPO_DIR = "SpotifyUnwrapped"
 
-# check if command exists
+# Function to check if a command exists
 function CommandExists {
-	param (
-		[string]$command
-		)
-	return Get-Command $command -ErrorAction SilentlyContinue
+    param (
+        [string]$command
+    )
+    return Get-Command $command -ErrorAction SilentlyContinue
 }
 
-# clone repo
+# Step 1: Clone the repository
 if (-not (Test-Path -Path $REPO_DIR)) {
-	get clone $REPO_URL $REPO_DIR
+    git clone $REPO_URL $REPO_DIR
 }
 
 Set-Location -Path $REPO_DIR
 
-# check if Python 3.12 is installed, install if not
+# Step 2: Check if Python 3.12 is installed and install it if not
 if (-not (CommandExists "python3.12")) {
-	Write-Host "Python 3.12 is not installed. Installing Python 3.12..."
-	$pythonInstaller = "https://www.python.org/ftp/python/3.12.0/python-3.12.0-amd64.exe"
-	$installerPath = "$env:TEMP\python-3.12.0-amd64.exe"
-	Invoke-WebRequest -Uri $pythonInstaller -OutFile $installerPath
-	Start-Process -Wait -FilePath $installerPath -ArgumentList "\quiet InstallAllUsers=1 PrependPath=1"
-	Remove-Item -Path $installer Path
+    Write-Host "Python 3.12 is not installed. Installing Python 3.12..."
+    $pythonInstaller = "https://www.python.org/ftp/python/3.12.0/python-3.12.0-amd64.exe"
+    $installerPath = "$env:TEMP\python-3.12.0-amd64.exe"
+    Invoke-WebRequest -Uri $pythonInstaller -OutFile $installerPath
+    Start-Process -Wait -FilePath $installerPath -ArgumentList "/quiet InstallAllUsers=1 PrependPath=1"
+    Remove-Item -Path $installerPath
 }
 
-# set up venv
+# Verify that python3.12 was installed correctly
+if (-not (CommandExists "python3.12")) {
+    Write-Host "Failed to install Python 3.12. Please install it manually and rerun the script."
+    exit 1
+}
+
+# Step 3: Set up a virtual environment
 if (-not (Test-Path -Path "venv")) {
-	python3.12 -m venv venv
+    python -m venv venv
 }
 
+# Step 4: Activate the virtual environment
 & .\venv\Scripts\Activate
 
+# Step 5: Install dependencies
 pip install -r requirements.txt
 
-# script to run analysis
+# Step 6: Create a script to run the analysis
 $scriptContent = @"
 @echo off
 call venv\Scripts\activate
-python spotifyAnalysis.py
+python spotify_analysis.py
 call venv\Scripts\deactivate
 "@
 Set-Content -Path "run_analysis.bat" -Value $scriptContent
 
-Write-Host "Setup is complete. To run analysis, place JSON files in the 'Spotify_Extended_Streaming_History' directory and run './run_analysis.sh' "
+Write-Host "Setup is complete. To run the analysis, place your JSON files in the 'Spotify_Extended_Streaming_History' directory and run 'run_analysis.bat'"
